requirements.txt
streamlit
pillow
transformers
torch
google-generativeai
